all tables are designed. testing is remaining
